backend/requirements.txt
# CivicLens Backend Dependencies

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# HTTP client
httpx>=0.27.0
requests>=2.31.0

# Web scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF extraction
PyMuPDF>=1.24.0
PyPDF2>=3.0.0

# Task scheduling
APScheduler>=3.10.0

# Embeddings (BAAI/bge-large-en-v1.5)
sentence-transformers>=3.0.0
numpy>=1.24.0

# Vector store (local fallback)
faiss-cpu>=1.7.4

# Vector store (cloud)
pinecone>=3.0.0

# Settings management
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Token counting
tiktoken>=0.6.0

# ScaleDown optional dependencies (semantic optimizer)
# Install from parent dir: pip install -e ../ or pip install -e ../[semantic]
